minor typing fix
